added document root for local url in getimagesize
discovered a problem with local URLs (relative to webroot) when adding marker icons, where the getimagesize was failing on a path like '/img/myicon.png'.  some digging on the web revealed that getimagesize wants the server's document root prepended to the url for the size operation to work, so that is what i added.  this made the complaint go away about finding the site-relative image, and the image shows as expected.